requirements.txt
Flask
Flask-SocketIO
eventlet
Flask-SQLAlchemy
python-dotenv
openai
numpy
psycopg2-binary